Move libcontainerd/shimopts to daemon/internal/libcontainerd/shimopts
