+++ daemon/runtime_unix.go
@@ -19,8 +19,8 @@ import (
 	"github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/config"
+	"github.com/docker/docker/daemon/internal/libcontainerd/shimopts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/libcontainerd/shimopts"
 	"github.com/moby/sys/atomicwriter"
 	"github.com/opencontainers/runtime-spec/specs-go/features"
 	"github.com/pkg/errors"
